Fixes to undoing/redoing and a few other tweaks

+++ KabasujiSinon/src/sinon/controllers/LevelStartController.java
@@ -6,7 +6,6 @@ import java.awt.event.ActionListener;
 import sinon.main.Game;
 import sinon.models.LightningLevel;
 import sinon.models.data.LevelType.Types;
-
 import sinon.views.LevelTypeInfoView;
 import sinon.views.LightningInfoView;
 import sinon.views.PuzzleInfoView;
@@ -53,16 +52,19 @@ public class LevelStartController implements ActionListener {
 		Types thisLevelsType = this.game.getLevel(this.levelNum).getLevelData().getLevelType();
 
 		switch (thisLevelsType) {
-			case PUZZLE:
-				lvlTypeInfoView = new PuzzleInfoView((int)this.game.getLevel().getLevelData().getLevelProperty().getMaxMoves());
-				break;
-			case LIGHTNING:
-				lvlTypeInfoView = new LightningInfoView(this.game.getLevel().getLevelData().getLevelProperty().getMaxTime(), (LightningLevel)this.game.getLevel(this.levelNum));
-				LightningTimerController l = new LightningTimerController((LightningLevel)this.game.getLevel(this.levelNum), (LightningInfoView)lvlTypeInfoView);
-				break;
-			case RELEASE:
-				lvlTypeInfoView = new ReleaseInfoView(this.game.getLevel().getLevelData().getLevelProperty().getReleaseSet());
-				break;
+		case PUZZLE:
+			lvlTypeInfoView = new PuzzleInfoView(this.game.getLevel().getLevelData().getLevelProperty().getMaxMoves());
+			break;
+		case LIGHTNING:
+			lvlTypeInfoView = new LightningInfoView(this.game.getLevel().getLevelData().getLevelProperty().getMaxTime(),
+					(LightningLevel) this.game.getLevel(this.levelNum));
+			LightningTimerController l = new LightningTimerController(
+					(LightningLevel) this.game.getLevel(this.levelNum), (LightningInfoView) lvlTypeInfoView);
+			break;
+		case RELEASE:
+			lvlTypeInfoView = new ReleaseInfoView(
+					this.game.getLevel().getLevelData().getLevelProperty().getReleaseSet());
+			break;
 		}
 
 		this.game.initializeMainView(this.levelSelectView, lvlTypeInfoView);
